pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermalright-lcd-control"
version = "1.3.0"
description = "Linux Thermal rigth lcd display control"
authors = [{ name = "REJEB BEN REJEB", email = "benrejebrejeb@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "PySide6>=6.5",
    "hid~=1.0.8",
    "psutil>=5.8.0",
    "opencv-python>=4.12.0.88",
    "pyusb>=1.3.1",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2"
]

[project.scripts]
application-name = "thermalright_lcd_control.main:main"